move valueOf function to ClassName.mockative.kt file
